pyproject.toml
[tool.poetry]
name = "bt_keyboard_adapter_pi"
version = "0.1.0"
description = ""
authors = ["kaiiiz <at881005@gmail.com>"]
packages = [
    { include = "bt_keyboard_adapter_pi" }
]

[tool.poetry.dependencies]
python = "^3.7"
pyserial = "^3.4"
keyboard = "^0.13.5"

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
